pyproject.toml
[tool.poetry]
name = "aw-watcher-afk-barrier"
version = "0.2.0"
description = ""
authors = ["Erik Bjäreholt <erik@bjareho.lt>", "Matas Minelga <minematas@gmail.com>"]
license = "MPL-2.0"

[tool.poetry.scripts]
aw-watcher-afk-barrier = "aw_watcher_afk-barrier:main"

[tool.poetry.dependencies]
python = "^3.7"
aw-client = "^0.5"
pyobjc-framework-Quartz = { version = "^7.0", platform = "darwin" }
python-xlib = { version = "^0.28", platform = "linux" }
pynput = "^1.7.3"

[tool.poetry.dev-dependencies]
mypy = "*"
macholib = {version = "^1.13", platform = "darwin"}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
